Move Page to navigator.py
