+++ miru/ext/nav/items.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import abc
-import attr
 import typing as t
 
 import hikari
@@ -276,49 +275,6 @@ class StopButton(NavButton):
         elif self.view.message:
             await self.view.message.edit(components=self.view)
         self.view.stop()
-
-
-@attr.define(slots=True)
-class Page:
-    """Allows for the building of more complex pages for use with NavigatorView."""
-    
-    content: hikari.UndefinedOr[t.Any] = hikari.UNDEFINED
-    """The content of the message. Anything passed here will be cast to str."""
-    attachment: hikari.UndefinedOr[hikari.Resourceish] = hikari.UNDEFINED
-    """An attachment to add to this page."""
-    attachments: hikari.UndefinedOr[t.Sequence[hikari.Resourceish]] = hikari.UNDEFINED
-    """A sequence of attachments to add to this page."""
-    embed: hikari.UndefinedOr[hikari.Embed] = hikari.UNDEFINED
-    """An embed to add to this page."""
-    embeds: hikari.UndefinedOr[t.Sequence[hikari.Embed]] = hikari.UNDEFINED
-    """A sequence of embeds to add to this page."""
-    mentions_everyone: hikari.UndefinedOr[bool] = hikari.UNDEFINED
-    """If True, mentioning @everyone will be allowed in this page's message."""
-    user_mentions: hikari.UndefinedOr[
-        t.Union[hikari.SnowflakeishSequence[hikari.PartialUser], bool]
-    ] = hikari.UNDEFINED
-    """The set of allowed user mentions in this page's message. Set to True to allow all."""
-    role_mentions: hikari.UndefinedOr[
-        t.Union[hikari.SnowflakeishSequence[hikari.PartialRole], bool]
-    ] = hikari.UNDEFINED
-    """The set of allowed role mentions in this page's message. Set to True to allow all."""
-
-    def _build_payload(self) -> dict[str, t.Any]:
-        d: dict[str, t.Any] = dict(
-            content=self.content or None,
-            attachments=self.attachments or None,
-            embeds=self.embeds or None,
-            mentions_everyone=self.mentions_everyone or False,
-            user_mentions=self.user_mentions or False,
-            role_mentions=self.role_mentions or False,
-        )
-        if not d["attachments"] and self.attachment:
-            d["attachments"] = [self.attachment]
-
-        if not d["embeds"] and self.embed:
-            d["embeds"] = [self.embed]
-        
-        return d
 
 
 # MIT License
